+++ guide/samples/src/lib.rs
@@ -18,6 +18,5 @@
 pub type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
 
 pub mod lro;
-pub mod mocking;
 pub mod polling_policies;
 pub mod retry_policies;
